rename reusable components and refactor them; add discussion post content validation
