+++ components/post_detail_components/replyMain.js
@@ -3,8 +3,8 @@ import useSWR from "swr";
 import {API_url} from "../../app_config";
 import {convertFromRaw, EditorState} from "draft-js";
 import CommentHeader from "./commentHeader";
-import DiscussionAddReply from "./discussionAddReply";
-import DiscussionCommentBox from "./discussionCommentBox";
+import AddCommentBox from "./addCommentBox";
+import CommentBox from './commentBox'
 
 export default function DiscussionReplyMain(props) {
     var postComments = []
@@ -35,8 +35,8 @@ export default function DiscussionReplyMain(props) {
     return (
         <>
             <CommentHeader commentLength={postComments.length}/>
-            <DiscussionAddReply postID={props.postID}/>
-            <DiscussionCommentBox replies={postComments} postID={props.postID}/>
+            <AddCommentBox postID={props.postID}/>
+            <CommentBox replies={postComments} postID={props.postID}/>
         </>
     )
 }
