+++ components/post_detail_components/upVoteSVG.js
