+++ pages/index.js
@@ -4,7 +4,7 @@ import { Col, Container, Row } from "react-bootstrap"
 import MyCalendar from "../components/calendar";
 import TodoBox from "../components/todo/todoBox";
 import AnnouncementBox from "../components/announcementBox";
-import NewsBox from "../components/newsBox";
+import NewsBox from "../components/campus_news/newsBox";
 import QuoteBox from "../components/quoteBox";
 import DDLBox from "../components/ddlBox";
 import useSWR from 'swr';
